first commita

+++ app/api/follow/route.js
@@ -1,5 +1,4 @@
 // url : http://localhost:3000/api/follow
-
 import { db } from "@/lib/db";
 import { NextResponse } from "next/server";
 
@@ -31,28 +30,35 @@ export const POST = async (request) => {
 };
 
 export const DELETE = async (request) => {
-  try {
-    // استخراج معرف العنصر المراد حذفه من الطلب
-    const { followingUsername, followerUsername } = await request.json();
+  if (request.method === "DELETE") {
+    try {
+      // استخراج معرف العنصر المراد حذفه من الجسم
+      const { followingUsername, followerUsername } = await request.json();
 
-    // حذف العنصر باستخدام Prisma Client
-    const deletedExample = await prisma.follow.delete({
-      where: {
-        followerUsername,
-        followingUsername,
-      },
-    });
+      // حذف العنصر باستخدام Prisma Client
+      const deletedExample = await prisma.follow.delete({
+        where: {
+          followerUsername,
+          followingUsername,
+        },
+      });
 
-    // إرجاع رسالة تأكيد مع البيانات المحذوفة
-    return NextResponse.json({
-      message: "Example deleted successfully",
-      data: deletedExample,
-    });
-  } catch (error) {
-    // إرجاع رسالة خطأ مفصلة في حالة حدوث خطأ
+      // إرجاع رسالة تأكيد مع البيانات المحذوفة
+      return NextResponse.json({
+        message: "Follow deleted successfully",
+        data: deletedExample,
+      });
+    } catch (error) {
+      // إرجاع رسالة خطأ مفصلة في حالة حدوث خطأ
+      return NextResponse.json(
+        { message: "Failed to delete follow", error },
+        { status: 500 }
+      );
+    }
+  } else {
     return NextResponse.json(
-      { message: "Failed to delete example", error },
-      { status: 500 }
+      { message: "Method Not Allowed" },
+      { status: 405 }
     );
   }
 };
